+++ app/models/document_models.py
@@ -29,6 +29,7 @@ class Document(Base):
 
     thumbnail_filename = Column(String(128), nullable=True, unique=True)
     comments_count = Column(Integer, index=True, default=0)
+    downloads_count = Column(Integer, index=True, default=0)
 
     document_user = relationship(
         "User", back_populates="user_documents", lazy="joined")
@@ -37,8 +38,13 @@ class Document(Base):
     document_tags = relationship(
         "Tag", back_populates="tag_document", lazy="joined",
         cascade="all, delete-orphan")
+
     document_comments = relationship(
         "Comment", back_populates="comment_document", lazy="noload",
+        cascade="all, delete-orphan")
+
+    document_downloads = relationship(
+        "Download", back_populates="download_document", lazy="noload",
         cascade="all, delete-orphan")
 
     def __init__(self, user_id: int, collection_id: int, document_name: str,
@@ -56,6 +62,7 @@ class Document(Base):
 
         self.thumbnail_filename = thumbnail_filename
         self.comments_count = 0
+        self.downloads_count = 0
 
     @property
     def file_path(self):
@@ -88,5 +95,6 @@ class Document(Base):
 
             "thumbnail_url": self.thumbnail_url,
             "comments_count": self.comments_count,
+            "downloads_count": self.downloads_count,
             "document_tags": self.tag_values,
         }